+++ march_hardware/test/mocks/mock_imotioncube.h
@@ -28,8 +28,8 @@ public:
   MOCK_METHOD0(getVelocityRadAbsolute, double());
 
   MOCK_METHOD0(getIMCVoltage, float());
-  MOCK_METHOD0(getMotorVoltage, float());
-  MOCK_METHOD0(getMotorCurrent, float());
+  MOCK_METHOD0(getMotorVoltage, double());
+  MOCK_METHOD0(getMotorCurrent, double());
 
   MOCK_METHOD1(actuateRad, void(double));
   MOCK_METHOD1(actuateTorque, void(int16_t));
